myproject/requirements.txt
Django==3.2.10
psycopg2-binary==2.9.1
openpyxl==3.0.9
pandas
psycopg2>=2.8.0
djangorestframework>=3.11.0
